// packages/extraction-contract/src/generated/annotations.ts
export const IFC_LOCATOR_KEYS = [
  "ifc",
  "globalId",
  "globalIds",
  "guid",
  "stepId",
  "stepIds",
  "step",
  "path"
] as const;